+++ users/views.go
@@ -89,7 +89,12 @@ func DeleteUser(response http.ResponseWriter, request *http.Request) {
 	defer db.Close()
 
 	var user User
-	db.Find(&user, "id = ?", params["id"])
+	db_error := db.Find(&user, "id = ?", params["id"])
+	if db_error.Error != nil {
+		response.WriteHeader(http.StatusNotFound)
+		json.NewEncoder(response).Encode(map[string]string{"detail": "User Not Found"})
+		return
+	}
 	db.Delete(&user)
 
 	response.WriteHeader(http.StatusNoContent)
@@ -107,7 +112,12 @@ func UpdateUser(response http.ResponseWriter, request *http.Request) {
 	defer db.Close()
 
 	var user User
-	db.Find(&user, "id = ?", params["id"])
+	db_error := db.Find(&user, "id = ?", params["id"])
+	if db_error.Error != nil {
+		response.WriteHeader(http.StatusNotFound)
+		json.NewEncoder(response).Encode(map[string]string{"detail": "User Not Found"})
+		return
+	}
 
 	err = json.NewDecoder(request.Body).Decode(&user)
 	if err != nil {
